pyproject.toml
[project]
name = "glabmetrics"
version = "1.0.0"
description = "Comprehensive CLI tool for GitLab instance analysis with detailed HTML reporting and optimization recommendations"
authors = [{name = "CDDS AB", email = "pypi@cdds.se"}]
maintainers = [{name = "CDDS AB", email = "pypi@cdds.se"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Environment :: Console",
]
keywords = [
    "gitlab", "devops", "analytics", "reporting", "cli", "statistics",
    "monitoring", "optimization", "git", "repository-management", "metrics"
]
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/cdds-ab/glabmetrics"
Documentation = "https://github.com/cdds-ab/glabmetrics#readme"
Repository = "https://github.com/cdds-ab/glabmetrics.git"
"Bug Tracker" = "https://github.com/cdds-ab/glabmetrics/issues"
Changelog = "https://github.com/cdds-ab/glabmetrics/blob/main/CHANGELOG.md"

[project.scripts]
glabmetrics = "glabmetrics.main:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
# Testing
test = "pytest"
test-cov = "pytest --cov=glabmetrics --cov-report=html:target/coverage-html --cov-report=term --cov-report=xml:target/coverage.xml"
test-fast = "pytest -x --tb=short"
test-verbose = "pytest -v --tb=long"

# Code Quality
lint = "flake8 glabmetrics tests"
format = { composite = ["black glabmetrics tests", "isort glabmetrics tests"] }
format-check = { composite = ["black --check glabmetrics tests", "isort --check-only glabmetrics tests"] }
typecheck = "mypy glabmetrics"
security = "bandit -r glabmetrics"
security-report = "bandit -r glabmetrics -f json -o target/bandit-report.json"
security-precommit = { shell = "pdm run bandit -r glabmetrics -ll -f json -o target/bandit-report.json; if [ $? -eq 1 ]; then echo '⚠️  Bandit found medium/high severity issues - commit blocked!'; exit 1; elif [ $? -eq 0 ]; then echo '✅ Bandit security scan passed'; exit 0; else echo '⚠️  Low severity security findings detected (warnings only)'; exit 0; fi" }

# Comprehensive CI-like checks
check = { composite = [
    "format-check",
    "lint",
    "typecheck",
    "security",
    "test-cov"
]}
check-fast = { composite = [
    "format-check",
    "lint",
    "test-fast"
]}

# Development utilities
clean = { shell = "rm -rf target/ .generated/ .pytest_cache/ .coverage htmlcov/ *.egg-info/" }
setup-dev = { shell = "rm -rf target/ .generated/ .pytest_cache/ .coverage htmlcov/ *.egg-info/ && mkdir -p target && pdm run pre-commit-install" }

# Release management
commit = "cz commit"
bump = "cz bump"
changelog = "cz changelog"
pre-commit-install = "pre-commit install"

# GitLab analysis commands
glabmetrics = "python -m glabmetrics.main"
analyze = "python -m glabmetrics.main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--color=yes"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "api: marks tests that test API interactions"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["glabmetrics"]
omit = [
    "tests/*",
    "glabmetrics/__main__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = false
version = "1.0.0"
version_files = [
    "glabmetrics/__init__.py:__version__"
]

# PDM versioning managed by commitizen

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.venv
    | venv
    | build
    | dist
    | target
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [".git", "__pycache__", ".venv", "venv", "env", "build", "dist", "*.egg-info", "htmlcov", "target"]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
no_strict_optional = true
check_untyped_defs = false
disallow_untyped_defs = false
# Disable strict checks for existing codebase
disable_error_code = ["var-annotated", "no-any-return", "attr-defined", "call-arg", "assignment", "arg-type", "return-value", "operator", "index", "misc"]
[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
    "freezegun>=1.2.0",
    "beautifulsoup4>=4.12.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "flake8-pyproject>=1.2.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0",
    "types-PyYAML>=6.0.0",
    "commitizen>=3.12.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.8.6",
]
